requirements.txt
spacy
inflect
numpy
pytracery
cairosvg
